chore: decouple each schema into separate field template
