+++ app/form-components/FormTemplate.tsx
@@ -1,28 +1,25 @@
 import React, { useState, useEffect, useCallback } from 'react';
+import { useRouter } from 'next/router';
+import styled from 'styled-components';
+import { isNil, throttle, padStart } from 'lodash';
 import FormHeader from 'form-components/FormHeader';
 import FormStage from 'form-components/FormStage';
 import Form from 'form-components/GovForm';
-import { getUISchema } from 'schemas-ui';
 import FormButtons from 'form-components/FormButtons';
-import { Request } from 'interfaces/Request';
-import CenteredModal from 'components/CenteredModal';
-import { createRequest, updateRequest } from 'services/request';
+import FieldTemplate from 'form-components/FieldTemplate';
 import ArrayFieldTemplate from 'form-components/ArrayFieldTemplate';
-import FormReview from 'form-components/FormReview';
-import TermsAndConditions from 'components/TermsAndConditions';
-import { useRouter } from 'next/router';
-import styled from 'styled-components';
-import { isNil, throttle } from 'lodash';
+import CenteredModal from 'components/CenteredModal';
 import { validateForm, customValidate } from 'utils/validate';
-import { getFormStageInfo } from 'utils/helpers';
-import { stageTitlesUsingForms, stageTitlesReviewing, createTeamModalId } from 'utils/constants';
+import { parseError, getFormStageInfo } from 'utils/helpers';
 import { withTopAlert, TopAlert } from 'layout/TopAlert';
 import { getMyTeams, getAllowedTeams } from 'services/team';
+import { getUISchema } from 'schemas-ui';
+import { Request } from 'interfaces/Request';
 import { SaveMessage } from 'interfaces/form';
 import { Team, LoggedInUser } from 'interfaces/team';
 import Link from '@button-inc/bcgov-theme/Link';
-import CreateTeamForm from 'form-components/team-form/CreateTeamForm';
 import CancelConfirmModal from 'page-partials/edit-request/CancelConfirmModal';
+import { createRequest, updateRequest } from 'services/request';
 import getConfig from 'next/config';
 const { publicRuntimeConfig = {} } = getConfig() || {};
 const { enable_gold } = publicRuntimeConfig;
@@ -38,11 +35,6 @@ const HeaderContainer = styled.div`
   min-height: 150px;
 `;
 
-const ErrorText = styled.p`
-  color: #d94532;
-  font-weight: bold;
-`;
-
 interface Props {
   currentUser: LoggedInUser;
   request?: Request | undefined;
@@ -53,7 +45,6 @@ function FormTemplate({ currentUser, request, alert }: Props) {
   const router = useRouter();
   const { step } = router.query;
   const stage = step ? Number(step) : 0;
-
   const [formData, setFormData] = useState((request || {}) as Request);
   const [formStage, setFormStage] = useState(stage);
   const [loading, setLoading] = useState(false);
@@ -62,8 +53,7 @@ function FormTemplate({ currentUser, request, alert }: Props) {
   const [errors, setErrors] = useState<any>({});
   const [visited, setVisited] = useState<any>(request ? { '0': true } : {});
   const [teams, setTeams] = useState<Team[]>([]);
-  const [showAccountableError, setShowAccountableError] = useState(false);
-  const [schemas, setSchemas] = useState<any>();
+  const [schemas, setSchemas] = useState<any[]>([]);
   const [schema, setSchema] = useState<any>();
   const [stageTitle, setStageTitle] = useState<any>();
   const [stages, setStages] = useState<any>();
@@ -72,6 +62,7 @@ function FormTemplate({ currentUser, request, alert }: Props) {
   const isAdmin = currentUser.isAdmin || false;
 
   const showFormButtons = formStage !== 0 || formData.usesTeam || formData.projectLead;
+  const isLastStage = formStage === schemas.length - 1;
 
   const throttleUpdate = useCallback(
     throttle(
@@ -100,9 +91,6 @@ function FormTemplate({ currentUser, request, alert }: Props) {
     }
     if (showModal) {
       window.location.hash = 'info-modal';
-      setShowAccountableError(true);
-    } else {
-      setShowAccountableError(false);
     }
 
     throttleUpdate(e);
@@ -160,7 +148,7 @@ function FormTemplate({ currentUser, request, alert }: Props) {
     router.push({ pathname: redirectUrl });
   };
 
-  const uiSchema = getUISchema(request as Request);
+  const uiSchema = getUISchema({ integration: request as Request, isAdmin });
 
   const handleFormSubmit = async () => {
     setLoading(true);
@@ -208,32 +196,78 @@ function FormTemplate({ currentUser, request, alert }: Props) {
     setLoading(false);
   };
 
+  const handleSubmit = async () => {
+    try {
+      const [, err] = await updateRequest(formData, true);
+
+      if (err) {
+        alert.show({
+          variant: 'danger',
+          fadeOut: 10000,
+          closable: true,
+          content: parseError(err).message,
+        });
+      } else {
+        alert.show({
+          variant: 'success',
+          fadeOut: 10000,
+          closable: true,
+          content: `Request ID:${padStart(String(formData.id), 8, '0')} is successfully submitted!`,
+        });
+
+        router.push({
+          pathname: isAdmin ? '/admin-dashboard' : '/my-dashboard',
+          query: { id: formData.id },
+        });
+      }
+    } catch (err) {
+      console.error(err);
+    }
+  };
+
   const handleButtonSubmit = async () => {
     if (formStage === 0) {
       if (isNew) {
         visited[formStage] = true;
         setVisited(visited);
-        return;
       }
+    } else if (isLastStage) {
+      openConfirmModal();
+    } else {
+      const newStage = formStage + 1;
+      changeStep(newStage);
     }
-
-    const newStage = formStage + 1;
-    changeStep(newStage);
   };
 
-  const handleTeamCreate = async (teamId: number) => {
-    await loadTeams();
-    setFormData({ ...formData, usesTeam: true, teamId: String(teamId) });
-  };
+  const openConfirmModal = () => {
+    const formErrors = validateForm(formData, schemas);
+    if (Object.keys(formErrors).length > 0) {
+      alert.show({
+        variant: 'danger',
+        fadeOut: 10000,
+        closable: true,
+        content:
+          'There were errors with your submission. Please see the navigation tabs above for the form pages with errors.',
+      });
 
-  const handleModalClose = () => {
-    window.location.hash = '#';
+      setErrors(formErrors);
+    } else {
+      window.location.hash = 'confirmation-modal';
+    }
   };
 
   const backButton = isApplied ? <CancelConfirmModal onConfirm={handleBackClick} /> : null;
-  const backButtonText = isApplied ? 'Cancel' : 'Save and Close';
+  const buttonTexts = { continue: '', back: '' };
 
-  if (!schemas) return null;
+  if (isLastStage) {
+    buttonTexts.continue = isApplied ? 'Update' : 'Submit';
+    buttonTexts.back = isApplied ? 'Cancel' : 'Save and Close';
+  } else {
+    buttonTexts.continue = 'Next';
+    buttonTexts.back = isApplied ? 'Cancel' : 'Save and Close';
+  }
+
+  if (schemas.length === 0) return null;
 
   return (
     <>
@@ -255,82 +289,52 @@ function FormTemplate({ currentUser, request, alert }: Props) {
           for more information.
         </Description>
       </HeaderContainer>
-      {stageTitle === 'Terms and conditions' && <TermsAndConditions />}
-      {stageTitlesReviewing.includes(stageTitle) && (
-        <FormReview
-          formData={formData}
-          setFormData={setFormData}
-          setErrors={setErrors}
-          isAdmin={isAdmin}
-          teams={teams}
-        />
-      )}
-      {stageTitlesUsingForms.includes(stageTitle) && (
-        <Form
-          schema={schema}
-          uiSchema={uiSchema}
-          onChange={handleChange}
-          onSubmit={handleFormSubmit}
-          formData={formData}
-          formContext={{ isAdmin, teams }}
-          ArrayFieldTemplate={ArrayFieldTemplate}
-          liveValidate={visited[formStage] || isApplied}
-          validate={customValidate}
-        >
-          {showFormButtons ? (
-            <FormButtons
-              formSubmission={formStage === 0}
-              backButton={backButton}
-              text={{ continue: 'Next', back: backButtonText }}
-              loading={loading}
-              handleSubmit={handleButtonSubmit}
-              handleBackClick={handleBackClick}
-              saving={saving}
-              saveMessage={saveMessage}
-            />
-          ) : (
-            <></>
-          )}
-        </Form>
-      )}
-      {stageTitle === 'Requester Info' && (
-        <>
-          <CenteredModal
-            id="info-modal"
-            content={
-              <>
-                <p>If you are not accountable for this project, please refer this request to someone who is.</p>
-                <p> Only the person who is responsible for this project can submit the intergration request.</p>
-              </>
-            }
-            icon={false}
-            onConfirm={handleModalClose}
-            confirmText="Close"
-            showCancel={false}
-            title="Project Accountability"
-            buttonStyle="custom"
-            buttonAlign="center"
-            closable
+      <Form
+        schema={schema}
+        uiSchema={uiSchema}
+        onChange={handleChange}
+        onSubmit={handleFormSubmit}
+        formData={formData}
+        formContext={{ isAdmin, teams, formData, setFormData, loadTeams }}
+        FieldTemplate={FieldTemplate}
+        ArrayFieldTemplate={ArrayFieldTemplate}
+        liveValidate={visited[formStage] || isApplied}
+        validate={customValidate}
+      >
+        {showFormButtons ? (
+          <FormButtons
+            formSubmission={formStage === 0}
+            backButton={backButton}
+            text={buttonTexts}
+            loading={loading}
+            handleSubmit={handleButtonSubmit}
+            handleBackClick={handleBackClick}
+            saving={saving}
+            saveMessage={saveMessage}
           />
-          <CenteredModal
-            title="Create a new team"
-            icon={null}
-            id={createTeamModalId}
-            content={
-              <CreateTeamForm onSubmit={(teamId: number) => handleTeamCreate(teamId)} currentUser={currentUser} />
-            }
-            showCancel={false}
-            showConfirm={false}
-            closable
-          />
-          {showAccountableError && (
-            <ErrorText>
-              If you are not accountable for this request, please refer this request to someone who is. Only the
-              responsible person can submit the integration request.
-            </ErrorText>
-          )}
-        </>
-      )}
+        ) : (
+          <></>
+        )}
+      </Form>
+      <CenteredModal
+        id={`confirmation-modal`}
+        content={
+          <>
+            <p>Are you sure you're ready to submit your request?</p>
+            {!isAdmin && (
+              <p>
+                If you need to change anything after submitting your request, please contact our{' '}
+                <Link external href="https://chat.developer.gov.bc.ca/channel/sso/">
+                  #SSO channel
+                </Link>{' '}
+                or email <Link href="mailto:bcgov.sso@gov.bc.ca">bcgov.sso@gov.bc.ca</Link>
+              </p>
+            )}
+          </>
+        }
+        title="Submitting Request"
+        onConfirm={handleSubmit}
+      />
     </>
   );
 }
