+++ app/form-components/team-form/CreateTeamForm.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { useState, useContext } from 'react';
 import Input from '@button-inc/bcgov-theme/Input';
 import styled from 'styled-components';
 import { Button } from '@bcgov-sso/common-react-components';
@@ -8,6 +8,7 @@ import { User, LoggedInUser } from 'interfaces/team';
 import { withTopAlert, TopAlert } from 'layout/TopAlert';
 import ErrorText from 'components/ErrorText';
 import TeamMembersForm from './TeamMembersForm';
+import { SessionContext, SessionContextInterface } from 'pages/_app';
 
 const ButtonsContainer = styled.div`
   display: flex;
@@ -20,7 +21,6 @@ const ButtonsContainer = styled.div`
 
 interface Props {
   onSubmit: (teamId: number) => void;
-  currentUser: LoggedInUser;
   alert: TopAlert;
 }
 
@@ -35,7 +35,10 @@ const emptyUser: User = {
   id: new Date().getTime(),
 };
 
-function CreateTeamForm({ onSubmit, currentUser, alert }: Props) {
+function CreateTeamForm({ onSubmit, alert }: Props) {
+  const context = useContext<SessionContextInterface | null>(SessionContext);
+  const { session } = context || {};
+
   const [members, setMembers] = useState<User[]>([emptyUser]);
   const [teamName, setTeamName] = useState('');
   const [loading, setLoading] = useState(false);
@@ -101,7 +104,12 @@ function CreateTeamForm({ onSubmit, currentUser, alert }: Props) {
       {errors && errors.name && <ErrorText>{errors?.name}</ErrorText>}
       <br />
       <strong>Team Members</strong>
-      <TeamMembersForm errors={errors} members={members} setMembers={setMembers} currentUser={currentUser} />
+      <TeamMembersForm
+        errors={errors}
+        members={members}
+        setMembers={setMembers}
+        currentUser={session as LoggedInUser}
+      />
       <ButtonsContainer>
         <Button variant="secondary" onClick={handleCancel}>
           Cancel
