+++ app/schemas-ui/ui-gold.ts
@@ -1,59 +1,72 @@
 import { isNil } from 'lodash';
-import FieldTemplateNoTitle from 'form-components/FieldTemplateNoTitle';
-import FieldTemplateWithTitle from 'form-components/FieldTemplateWithTitle';
-import AddTeamWidget from 'form-components/AddTeamWidget';
-import ClientTypeWidget from 'form-components/widgets/ClientTypeWidget';
-import { Request } from 'interfaces/Request';
+import FieldProjectTeam from '@app/form-components/FieldProjectTeam';
+import ClientTypeWidget from '@app/form-components/widgets/ClientTypeWidget';
+import FieldTermsAndConditions from '@app/form-components/FieldTermsAndConditions';
+import FieldRequesterInfo from '@app/form-components/FieldRequesterInfo';
+import FieldReviewAndSubmit from '@app/form-components/FieldReviewAndSubmit';
+import { Request } from '@app/interfaces/Request';
 
-const getUISchema = (request: Request) => {
-  const isNew = isNil(request?.id);
-  const isApplied = request?.status === 'applied';
+interface Props {
+  integration: Request;
+  isAdmin: boolean;
+}
 
-  const envDisabled = isApplied ? request?.environments?.concat() || [] : ['dev'];
+const getUISchema = ({ integration, isAdmin }: Props) => {
+  const isNew = isNil(integration?.id);
+  const isApplied = integration?.status === 'applied';
+
+  const envDisabled = isApplied ? integration?.environments?.concat() || [] : ['dev'];
+  const includeComment = isApplied && isAdmin;
 
   return {
     projectName: {
-      'ui:FieldTemplate': FieldTemplateNoTitle,
       'ui:placeholder': 'Project Name',
+      'ui:label': false,
     },
     usesTeam: {
       'ui:widget': 'radio',
-      'ui:FieldTemplate': FieldTemplateWithTitle,
-      'ui:readonly': isApplied && request?.usesTeam,
+      'ui:readonly': isApplied && integration?.usesTeam,
     },
     projectLead: {
+      'ui:FieldTemplate': FieldRequesterInfo,
       'ui:widget': 'radio',
-      'ui:FieldTemplate': FieldTemplateWithTitle,
       'ui:readonly': !isNew,
     },
     newToSso: {
       'ui:widget': 'radio',
-      'ui:FieldTemplate': FieldTemplateWithTitle,
     },
     publicAccess: {
       'ui:widget': ClientTypeWidget,
-      'ui:FieldTemplate': FieldTemplateWithTitle,
     },
     devIdps: {
       'ui:widget': 'checkboxes',
     },
     bceidTo: {
-      'ui:FieldTemplate': FieldTemplateNoTitle,
+      'ui:label': false,
       'ui:readonly': true,
     },
     bceidCc: {
-      'ui:FieldTemplate': FieldTemplateNoTitle,
+      'ui:label': false,
     },
     bceidBody: {
       'ui:widget': 'textarea',
     },
     environments: {
       'ui:widget': 'checkboxes',
-      'ui:FieldTemplate': FieldTemplateWithTitle,
       'ui:enumDisabled': envDisabled,
     },
     createTeam: {
-      'ui:FieldTemplate': AddTeamWidget,
+      'ui:FieldTemplate': FieldProjectTeam,
+      'ui:widget': 'hidden',
+      'ui:label': false,
+    },
+    agreeWithTerms: {
+      'ui:FieldTemplate': FieldTermsAndConditions,
+    },
+    comment: {
+      'ui:FieldTemplate': FieldReviewAndSubmit,
+      'ui:widget': includeComment ? 'textarea' : 'hidden',
+      'ui:label': false,
     },
   };
 };
