+++ app/schemas/commentUi.ts
@@ -1,9 +1,7 @@
-import FieldTemplateNoTitle from 'form-components/FieldTemplateNoTitle';
-
 const commentUi = {
   comment: {
-    'ui:FieldTemplate': FieldTemplateNoTitle,
     'ui:widget': 'textarea',
+    'ui:label': false,
   },
 };
